% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setting_data.R
\name{ACE_bycol}
\alias{ACE_bycol}
\title{Applying ACE transformation by column}
\usage{
ACE_bycol(mat, kine, categ)
}
\arguments{
\item{mat}{sc or wf matrix}

\item{kine}{kinematic matrix, either training or testing, according to mat}

\item{categ}{indicates whether categorica variables are present,
0 for spike counts, NULL for waveforms}
}
\value{
ace transformed data
}
\description{
This function applyies ACE transformation by column for either SC or WF
}
