% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrodes_info_functions.R
\name{build_df_equations}
\alias{build_df_equations}
\title{Build the equation info dataframe}
\usage{
build_df_equations(r2, z_colnames, n_units, n_lags = 13, n_transf = 3,
  n_wf = 0, n_feat = 4, n_transf_wf = 1)
}
\arguments{
\item{r2}{r-square score}

\item{z_colnames}{names of the spike count matrix columns}

\item{n_units}{number of units in the data set}

\item{n_lags}{number of lags in the data set}

\item{n_transf}{number of spike count transformation in the data set}

\item{n_wf}{number of waveforms in the data set}

\item{n_feat}{number of waveform features in the data set}
}
\description{
This function builds the equation info dataframe, with info for every equation about electrode, lag, transformation, waveforms characteristics and r-square score
}
