% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setting_data.R
\name{build_wf_mat}
\alias{build_wf_mat}
\title{Building WF data eliminating dead channel}
\usage{
build_wf_mat(wf_mat, dead_channels)
}
\arguments{
\item{wf_mat}{waveform matrix}

\item{dead_channels}{channels to eliminate}
}
\value{
a list of a list, as in input, but without non-active channels
}
\description{
This function builds the waveform matrix, that by default should be a list
with 13 elements (lag), each element is a list of 4 matrixes (features),
each matrix is T x n_units
}
