% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{filter}
\alias{filter}
\title{Kalman Filter}
\usage{
filter(sp, yt, Xo, Xs, smooth = FALSE)
}
\arguments{
\item{sp}{list describing the state space model}

\item{yt}{matrix of data}

\item{Xo}{matrix of exogenous observation data}

\item{Xs}{matrix of exogenous state data}

\item{smooth}{boolean indication whether to run the backwards smoother}
}
\description{
Kalman Filter
}
