% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrodes_info_functions.R
\name{get_bias}
\alias{get_bias}
\title{Get tuning curves bias}
\usage{
get_bias(z, cov)
}
\arguments{
\item{z}{matrix of spike counts}

\item{cov}{matrix of covariates}
}
\description{
This function computes the bias for each tuning-curve model
}
