% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_model_functions.R
\name{get_current_model_info}
\alias{get_current_model_info}
\title{Computes the current error}
\usage{
get_current_model_info(vel = velocity.Train, veltest = velocity.Test, z,
  ztest, A = NULL, W = NULL, H = NULL, Q = NULL,
  int_train = int.Train, time_train = Time.Train, int_test = int.Test,
  time_test = Time.Test, encoding = TRUE, decoding = TRUE, mle = TRUE,
  decoding_train = TRUE, weights = NULL)
}
\arguments{
\item{vel}{velocity in the training set}

\item{veltest}{velocity in the testing set}

\item{z}{neural data in the training set}

\item{ztest}{neaural data in the testing set}

\item{A}{the state matrix}

\item{W}{the state covariance matrix}

\item{H}{the encoding matrix}

\item{Q}{the encoding covariance matrix}

\item{int_train}{vector of integer on when the trial starts in the training set}

\item{time_train}{vector of time of the intervals in the training set}

\item{int_test}{vector of integer on when the trial starts in the testing set}

\item{time_test}{vector of time of the intervals in the testing set}

\item{encoding}{whether to run the encoding model or use given A,W,H,Q}

\item{decoding}{whether to run the decoding model or extract only encoding pars}

\item{mle}{whether compute prediction for MLE or KF}

\item{decoding_train}{whether decoding on training set}

\item{weights}{set of weights for mle}
}
\value{
a list of values with encoding parameter and decoded mse
}
\description{
This function computes current error, given a certain model
}
