% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrodes_info_functions.R
\name{get_equations_info}
\alias{get_equations_info}
\title{Extracting equation info}
\usage{
get_equations_info(experiment_fname, z, n_units, n_lags, n_wf, n_feat, n_transf,
  save_name, vel, n_transf_wf = 1, compute_rsq = TRUE)
}
\arguments{
\item{experiment_fname}{experiment file name to be loaded}

\item{z}{spike count training set}

\item{n_units}{number of units in the data set}

\item{n_lags}{number of lags in the data set}

\item{n_wf}{number of waveforms in the data set}

\item{n_feat}{number of waveform features in the data set}

\item{n_transf}{number of spike count transformation in the data set}

\item{save_name}{name of the file where saving the equations info}

\item{vel}{velocity training set}
}
\description{
This function extracts equation info, and saves it in save_name
}
