% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short-cut_functions.R
\name{get_inverse_matrix_m1}
\alias{get_inverse_matrix_m1}
\title{Compute inverse of a symmetrix matrix, leaving out one row and column}
\usage{
get_inverse_matrix_m1(Qm1, ind)
}
\arguments{
\item{Qm1}{inverse of the current(full) symmetric matrix}

\item{ind}{index to leave out, or set of indeces to leave out}
}
\description{
This function computes the inverse of a symmetrix matrix, leaving out one row
and column using the short-cut formula provived in my thesis
}
