% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short-cut_functions.R
\name{get_inverse_matrix_p1}
\alias{get_inverse_matrix_p1}
\title{Compute inverse of a symmetrix matrix, adding one row and column}
\usage{
get_inverse_matrix_p1(Qi, Qm1)
}
\arguments{
\item{Qi}{full matrix composed of n+1 x n+1 element, the last one is
the one to add}

\item{Qm1}{inverse of the current used symmetric matrix}
}
\description{
This function computes the inverse of a symmetrix matrix, adding one row and column
using the short-cut formula provived in my thesis
}
