% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_model_functions.R
\name{get_mse_trial}
\alias{get_mse_trial}
\title{MSE on a single trial}
\usage{
get_mse_trial(predicted_vel, true_vel)
}
\arguments{
\item{predicted_vel}{centered predicted velocity}

\item{true_vel}{centered true velocity}
}
\value{
mse between true velocity and predicted velocity for a given trial
}
\description{
This function computes the mse between true velocity and predicted velocity on a single trial.
}
