% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrodes_info_functions.R
\name{get_optimal_boxcox}
\alias{get_optimal_boxcox}
\title{Get tuning curves optimal boxcox transformation}
\usage{
get_optimal_boxcox(z, cov, shift = 1e-04)
}
\arguments{
\item{z}{matrix of spike counts}

\item{cov}{matrix of covariates}

\item{shift}{shift for computing boxcox transformation for avoiding dividing by zero}
}
\description{
This function computes the optimal boxcox transformation for each tuning-curve model
}
