% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrodes_info_functions.R
\name{get_r2}
\alias{get_r2}
\title{Get tuning curves r-square}
\usage{
get_r2(z, cov)
}
\arguments{
\item{z}{matrix of spike counts}

\item{cov}{matrix of covariates}
}
\description{
This function computes r squared for each tuning-curve model
}
