% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrodes_info_functions.R
\name{get_sigma2}
\alias{get_sigma2}
\title{Get tuning curves sigma-square}
\usage{
get_sigma2(z, cov)
}
\arguments{
\item{z}{matrix of spike counts}

\item{cov}{matrix of covariates}
}
\description{
This function computes sigma square for each tuning-curve model
}
