% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_model_functions.R
\name{get_tot_mse}
\alias{get_tot_mse}
\title{MSE on the whole experiment}
\usage{
get_tot_mse(int_trial, time_trial, vtrue, vhat)
}
\arguments{
\item{int_trial}{interval of time, can be training or testing}

\item{time_trial}{Time of the experiment, can be training or testing}

\item{vtrue}{centered true velocity matrix}

\item{vhat}{centered decoded velocity matrix. Can be in a list form.}
}
\value{
mse for all trials in the training set
}
\description{
This function computes the mse between true velocity and predicted velocity on the whole experiment. Variables need to be previously centered
}
