% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setting_data.R
\name{get_transformed_scdata}
\alias{get_transformed_scdata}
\title{Transforming data}
\usage{
get_transformed_scdata(enc_mat, dec_mat, vel_train = velocity.Train,
  vel_test = velocity.Test, dead_channels = NULL,
  categorical_transf = TRUE)
}
\arguments{
\item{enc_mat}{the encoding matrix of data}

\item{dec_mat}{the decoding matrix of data}

\item{vel_train}{matrix of velocity from training set}

\item{vel_test}{matrix of velocity from testing set}

\item{dead_channels}{non active channels}
}
\description{
This function transformes data using sqrt and ACE transformation
}
