% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_filter_functions.R
\name{kalman_decode}
\alias{kalman_decode}
\title{Function to decode the Kalman Filter trajectory}
\usage{
kalman_decode(A = NULL, W = NULL, H = NULL, Q = NULL, z,
  initial_state = rep(0, ncol(H)), MLE = TRUE, weights = NULL,
  Mole = NULL)
}
\arguments{
\item{A}{state matrix}

\item{W}{state covariance matrix}

\item{H}{encoding matrix}

\item{Q}{encoding covariance matrix}

\item{z}{centered spike count matrix}

\item{initial_state}{centered initial velocity state}

\item{MLE}{boolean for whether or not decoding is done including a prior model, MLE=TRUE/FALSE exclude/includes the prior model}

\item{weights}{possible weights to weight the covariance matrix}

\item{Mole}{matrix needed only for computing the MLE}

\item{weights}{weights to rescale the covariance matrix}
}
\value{
decoded trajectory for a given trial if not MLE, for the whole testing
set if MLE
}
\description{
This function decodes trajectory for a state space model with state parameters A and W and observation parameters H and Q
}
