% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_filter_functions.R
\name{kalman_decode_aoo}
\alias{kalman_decode_aoo}
\title{Decoding trajectory with short cut formula for adding one equation in}
\usage{
kalman_decode_aoo(A = NULL, W = NULL, H = NULL, Q = NULL,
  length_current, z, initial_state = rep(0, ncol(H)))
}
\arguments{
\item{A}{state matrix}

\item{W}{state covariance matrix}

\item{H}{encoding matrix for the full model}

\item{Q}{encoding covariance matrix for the full model}

\item{length_current}{length of current model}

\item{z}{centered spike count matrix}

\item{initial_state}{centered initial velocity state}
}
\value{
a list of decoded trajectory for a given trial obtained by adding
one equation at the time. Every element of the list corresponds to an equation
added
}
\description{
Function to decode the trajectory using a short cut formula for adding
 one equation in.
}
