% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_filter_functions.R
\name{kalman_decode_loo}
\alias{kalman_decode_loo}
\title{Decoding trajectory with short cut formula for leaving one equation out}
\usage{
kalman_decode_loo(A = NULL, W = NULL, H = NULL, Q = NULL, z,
  length_current = NULL, initial_state = rep(0, ncol(H)))
}
\arguments{
\item{A}{the state matrixe}

\item{W}{the state covariance matrix}

\item{H}{the encoding matrix}

\item{Q}{the encoding covariance matrix}

\item{z}{the centered spike count matrix}

\item{length_current}{length of current model, needed only for add_one_in}

\item{initial_state}{is the centered initial velocity state}

\item{weights}{possible weights to weight the covariance matrix}
}
\value{
a list of decoded trajectory for a given trial obtained by dropping
one equation at the time. Every element of the list corresponds to an equation
dropped
}
\description{
Function to decode the trajectory using a short cut formula for leaving one equation out.
}
