% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_filter_functions.R
\name{kalman_train}
\alias{kalman_train}
\title{Function to encode parameters of a Kalman filter}
\usage{
kalman_train(x, z, interruptions = NA, fit_A = FALSE, noise_units = rep(0,
  ncol(z)), unit_interactions = TRUE, single_neuron = 0, lambda = 0)
}
\arguments{
\item{x}{state (usually T x 3)}

\item{z}{spike counts (usually T x 96)}

\item{interruptions}{NA unless the data is segmented (i.e. time line
is interrupted) A printed string}

\item{fit_A}{fit a transition matrix. fit_A = FALSE uses an identity state
transition matrix}

\item{noise_units}{forces flat tuning curves for a subset if 1}

\item{unit_interactions}{if TRUE the state transition var-covar matrix is
diagonal}

\item{single_neuron}{if = 0 all are considered, otherwise the numb of
specific neuron has to be specified}

\item{lambda}{parameter for ridge}
}
\value{
a list of estimated parameters for the state space model
}
\description{
This function encodes parameters for a state space model with state
parameters A and W and observation parameters H and Q
}
