% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{kalmanfilter}
\alias{kalmanfilter}
\title{Kalman Filter}
\description{
\emph{kalmanfilter} Rcpp implementation of the multivariate Kalman filter for 
state space models that can handle missing values and exogenous data in the 
observation and state equations. See the package vignette using 
\code{browseVignettes("kalmanfilter")} to view it in your browser.
}
\author{
Alex Hubbard
}
