% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_filter_functions.R
\name{kf_aoo_objects}
\alias{kf_aoo_objects}
\title{Prepare the add-one-in objects}
\usage{
kf_aoo_objects(Hfull, Qfull, Sm1, incl, z, index, t, P_p)
}
\arguments{
\item{Hfull}{H parameter-matrix with all equations to add}

\item{Qfull}{Q covarance-matrix with all equations to add}

\item{incl}{length of included equation}

\item{z}{spike count matrix}

\item{index}{index to include}

\item{t}{time step}

\item{P_p}{prior covariance matrix from KF formula}
}
\value{
a list with parameters rescaled according to the add-one-in formula
}
\description{
This function prepare the add-one-in objects for the kf update
}
