% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_filter_functions.R
\name{kf_initialize}
\alias{kf_initialize}
\title{Initialize kf objects}
\usage{
kf_initialize(d, TT, initial_state, n_row)
}
\arguments{
\item{d}{dimension of the prediction object, should be 3}

\item{TT}{length of the trial time}

\item{initial_state}{initial state for the kf}

\item{n_row}{how many element to initialize, should match number of equations
to leave out/add in}
}
\value{
a list of initialized
}
\description{
This function initialize the kf objects
}
