% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_filter_functions.R
\name{kf_loo_objects}
\alias{kf_loo_objects}
\title{Prepare the leave-one-out objects}
\usage{
kf_loo_objects(H, Sm1, z, index, t)
}
\arguments{
\item{H}{H parameter-matrix with all the equations}

\item{z}{spike count matrix}

\item{index}{index to exclude}

\item{t}{time step}

\item{Q}{Q covariance-matrix with all the equations}
}
\value{
a list with parameters rescaled according to the leave-one-out formula
}
\description{
This function prepare the add-one-in objects for the kf update
}
