% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_filter_functions.R
\name{kf_update_kernel}
\alias{kf_update_kernel}
\title{Updates kalman filter solution regardless leave-one-out/add-one-in implementation}
\usage{
kf_update_kernel(A, W, kf_obj_i, other_shortcut_obj, t, d)
}
\arguments{
\item{A}{prior model parameter matrix}

\item{W}{prior model covariance matrix}

\item{kf_obj_i}{kf objects at time t-1 for all the equations}

\item{other_shortcut_obj}{objects containing specific leave-one/add-one stuff}

\item{t}{time step}

\item{d}{dimension of the state}
}
\value{
a list with parameters rescaled according to the leave-one-out formula
}
\description{
This function updates kalman filter solution regardless
leave-one-out/add-one-in implementation
}
