% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{likelihood}
\alias{likelihood}
\title{Kalman Likelihood}
\usage{
likelihood(sp, yt, Xo, Xs, w)
}
\arguments{
\item{sp}{list describing the state space model}

\item{yt}{matrix of data}

\item{Xo}{matrix of exogenous observation data}

\item{Xs}{matrix of exogenous state data}

\item{w}{matrix of weights}
}
\description{
Kalman Likelihood
}
