% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setting_data.R
\name{select_movement}
\alias{select_movement}
\title{Select portion of movement}
\usage{
select_movement(int, how_long = 10)
}
\arguments{
\item{int}{vector of intervals}

\item{how_long}{how many steps to select, from 0 to how_long}
}
\description{
This function selects a portion of movement
}
