% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setting_data.R
\name{transform_wf_mat}
\alias{transform_wf_mat}
\title{Transform WF matrix}
\usage{
transform_wf_mat(wf_mat, FUN, kine = NULL)
}
\arguments{
\item{wf_mat}{waveform matrix}

\item{FUN}{function to use to modify the data}

\item{kine}{kinematic data, needed only for ace transformation}
}
\value{
a matrix with the same dimension of the original one, but transformed
data
}
\description{
This function transform the waveform matrix, that is now a matrix T x
n_units n_lags n_feat in matrix where each column has been transformed
according to the FUN function passed
}
