% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{treasuries}
\alias{treasuries}
\title{Treasuries}
\format{
data.table with columns DATE, VARIABLE, VALUE, and MATURITY 
The data is quarterly frequency with variables DGS1MO, DGS3MO, DGS6MO, DGS1, 
DGS2, DGS3, DGS5, DGS7, DGS10, DGS20, and DGS30
}
\source{
FRED
}
\usage{
data(treasuries)
}
\description{
Treasuries
}
\keyword{datasets}
