% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_model_functions.R
\name{vhat_fun}
\alias{vhat_fun}
\title{Decoded velocity}
\usage{
vhat_fun(Mole = NULL, int_trial, time_trial, z, v, mle = TRUE, A = NULL,
  W = NULL, H = NULL, Q = NULL)
}
\arguments{
\item{Mole}{matrix needed only for computing the MLE}

\item{int_trial}{interval of time, can be training or testing}

\item{time_trial}{Time of the experiment, can be training or testing}

\item{z}{centered spike count matrix}

\item{v}{centered velocity matrix}

\item{A}{state matrix}

\item{W}{state covariance matrix}

\item{H}{encoding matrix}

\item{Q}{encoding covariance matrix}

\item{MLE}{boolean for whether or not decoding is done including a prior model, MLE=TRUE/FALSE exclude/includes the prior model}
}
\value{
predicted velocity in either MLE or KF scenario
}
\description{
This function computes the decoded velocity. Variables needs to be centered
}
