% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalmanfilter-package.R
\docType{package}
\name{kalmanfilter-package}
\alias{kalmanfilter}
\alias{kalmanfilter-package}
\title{kalmanfilter: Kalman Filter}
\description{
'Rcpp' implementation of the multivariate Kalman filter for state space models that can handle missing values and exogenous data in the observation and state equations. Kim, Chang-Jin and Charles R. Nelson (1999) "State-Space Models with Regime Switching: Classical and Gibbs-Sampling Approaches with Applications" \doi{10.7551/mitpress/6444.001.0001}\url{http://econ.korea.ac.kr/~cjkim/}.
}
\author{
\strong{Maintainer}: Alex Hubbard \email{hubbard.alex@gmail.com}

}
\keyword{internal}
